1.0.10 1. Использование медианы для подсчета среднего балла за урок или работу 2. Улучшено содержание текстовых элементов на сайте 3. Добавлена ссылка на репозитории GitHub на сайте 4. Использование дробного статуса новой версии и отправка его клиенту вместе с текстовым

+++ src/models/version_model.py
@@ -1,5 +1,5 @@
+from sqlalchemy import Numeric, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import SmallInteger, Integer, String
 
 from .base_model import BaseModel
 
@@ -10,7 +10,7 @@ __all__ = ['Version']
 class Version(BaseModel):
     number: Mapped[int] = mapped_column(Integer, primary_key=True)
     version: Mapped[str] = mapped_column(String(16))
-    status_id: Mapped[int] = mapped_column(SmallInteger)
+    status_id: Mapped[float] = mapped_column(Numeric)
     status: Mapped[str] = mapped_column(String(32))
     logs: Mapped[str] = mapped_column(String)
     date: Mapped[str] = mapped_column(String(16))
